Add  seo  comments

+++ resources/views/n_e_w_s/table.blade.php
@@ -2,8 +2,9 @@
 	 <table id="example" class="display" style="width:100%">
         <thead>
             <tr>
+            <th>  post</th>
             <th>  Photo</th>
-		              <th >created </th>
+		              <th >cat  </th>
 
              <th  >Action</th>
             </tr>
@@ -11,8 +12,24 @@
        <tbody>
     @foreach($nEWS as $nEWS)
         <tr>
+  			              <td>  
+						  @foreach($nEWS->get_NEWS_description   as $description )
+							   {{ $description->title}}
+							 @endforeach
+
+ </td>
   			              <td>     <img src="{{ URL::to('/').'/images/'.$nEWS->single_photo}}"  width="50" height="50">  </td>
-            <td>{!! $nEWS->created_at !!}</td>
+            <td> 
+  @foreach($categories_news    as $categories_news_val )
+  @foreach($categories_news_val->get_categories_news_ar_description   as $description )
+  <?php if (  $nEWS->get_cat['id']== $categories_news_val->id ){echo" $description->title";} ?>
+						 
+							   
+							 @endforeach
+							 @endforeach
+
+ 			
+							 </td>
 
              <td>
                 {!! Form::open(['route' => ['nEWS.destroy', $nEWS->id], 'method' => 'delete']) !!}
